Refactored metrics management: Expose metrics to embedded server
