+++ broker/src/main/java/io/moquette/server/netty/metrics/BytesMetricsHandler.java
@@ -26,6 +26,9 @@ import io.netty.util.AttributeKey;
 
 import static io.netty.channel.ChannelFutureListener.CLOSE_ON_FAILURE;
 
+import io.moquette.server.metrics.BytesMetrics;
+import io.moquette.server.metrics.BytesMetricsCollector;
+
 public class BytesMetricsHandler extends ChannelDuplexHandler {
 
     private static final AttributeKey<BytesMetrics> ATTR_KEY_METRICS = AttributeKey.valueOf("BytesMetrics");
